Add basic Makefile for compiling/running playc_tests
Right now the Makefile simply identifies the tests and compiles them
with the necessary dependencies. Work remains to be done to actually
make the tests compile (valid package level as3 syntax is not valid
in PlayScript).

The Makefile itself is also not complete, b/c of the following:
- it needs to be updated to fit into the top level mono Makefile
structure/convention
- it includes hard coded paths